+++ src/main/java/heavysnow/heath/dto/comment/CommentCreateResponseDto.java
@@ -1,4 +1,4 @@
-package heavysnow.heath.dto;
+package heavysnow.heath.dto.comment;
 
 
 import lombok.AccessLevel;
