+++ src/main/java/heavysnow/heath/service/CommentService.java
@@ -3,16 +3,15 @@ package heavysnow.heath.service;
 import heavysnow.heath.domain.Comment;
 import heavysnow.heath.domain.Member;
 import heavysnow.heath.domain.Post;
-import heavysnow.heath.dto.CommentCreateDto;
-import heavysnow.heath.dto.CommentCreateResponseDto;
-import heavysnow.heath.dto.CommentUpdateDto;
+import heavysnow.heath.dto.comment.CommentCreateDto;
+import heavysnow.heath.dto.comment.CommentCreateResponseDto;
+import heavysnow.heath.dto.comment.CommentUpdateDto;
 import heavysnow.heath.exception.BadRequestException;
 import heavysnow.heath.exception.ForbiddenException;
 import heavysnow.heath.exception.NotFoundException;
 import heavysnow.heath.repository.CommentRepository;
 import heavysnow.heath.repository.MemberRepository;
 import heavysnow.heath.repository.PostRepository;
-import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
